+++ src/main/java/cech12/extendedmushrooms/block/mushrooms/AmethystDeceiver.java
@@ -1,0 +1,17 @@
+package cech12.extendedmushrooms.block.mushrooms;
+
+import cech12.extendedmushrooms.init.ModFeatures;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+
+import javax.annotation.Nonnull;
+
+public class AmethystDeceiver extends BigMushroom {
+
+    @Nonnull
+    @Override
+    public ResourceKey<ConfiguredFeature<?, ?>> getBigMushroomFeature() {
+        return ModFeatures.BIG_AMETHYST_DECEIVER_CONFIGURED;
+    }
+
+}